requirements.txt
anthropic>=0.25.0
flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0
pandas>=2.1.0
numpy>=1.26.0
spacy>=3.7.0
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
biopython>=1.83
python-dotenv>=1.0.0
scikit-learn>=1.4.0
tqdm>=4.66.0
lxml>=5.1.0
pyngrok>=7.0.0
